made changes
The updated `page.svelte` code includes an "Add Node" button that allows you to add new nodes to the mind map, each of which can be dragged and repositioned. Additionally, a "Connect Nodes" button lets you select two nodes by clicking on them and then connect them with a line, with the lines dynamically updating as the nodes are moved.